created crud api in bankdetail of school, family, company in admin side

+++ database/migrations/2024_01_03_122905_add_extra_column_in_payments_table.php
@@ -14,7 +14,7 @@ class AddExtraColumnInPaymentsTable extends Migration
     public function up()
     {
         Schema::table('payments', function (Blueprint $table) {
-            $table->unsignedInteger('bank_detail_id')->nullable();
+            $table->unsignedBigInteger('bank_detail_id')->nullable();
             $table->foreign('bank_detail_id')->references('id')->on('bank_details')->onDelete('cascade'); 
         });
     }
